[CINN]delete auto-schedule
