--- paddle/cinn/auto_schedule/analysis/CMakeLists.txt
@@ -1,3 +1,0 @@
-core_gather_headers()
-
-gather_srcs(cinnapi_src SRCS analyze_ir.cc)
