--- paddle/cinn/auto_schedule/auto_schedule.proto
@@ -1,26 +1,0 @@
-// Copyright (c) 2022 CINN Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-syntax ="proto3";
-
-package cinn.auto_schedule.proto;
-
-import "paddle/cinn/ir/schedule/schedule_desc.proto";
-
-message TuningRecord {
-  string task_key = 1;
-  double execution_cost = 2;
-  double predicted_cost = 3;
-  cinn.ir.proto.ScheduleDesc trace = 4;
-}
