--- paddle/cinn/auto_schedule/search_space/CMakeLists.txt
@@ -1,5 +1,0 @@
-add_subdirectory(auto_gen_rule)
-
-core_gather_headers()
-
-gather_srcs(cinnapi_src SRCS search_state.cc)
