--- paddle/cinn/auto_schedule/search_space/auto_gen_rule/CMakeLists.txt
@@ -1,3 +1,0 @@
-core_gather_headers()
-
-gather_srcs(cinnapi_src SRCS auto_gen_rule.cc auto_inline.cc)
